Move EditorButton to a separate file
